feat: Add modal window to homepage for important information
The `App.jsx` file in the `client/src/pages` directory has been modified to add a modal window to the homepage. This modal window can be used to display important information to the user. The changes include adding a `Modal` component from the `@mantine/core` library and a button to open and close the modal. This enhancement improves the user experience by providing a visually appealing and interactive way to present important information on the homepage.